+++ DependencyInjection/Extension.php
@@ -22,33 +22,4 @@ use Symfony\Component\DependencyInjection\Extension\Extension as BaseExtension;
  */
 abstract class Extension extends BaseExtension
 {
-    private array $annotatedClasses = [];
-
-    /**
-     * Gets the annotated classes to cache.
-     *
-     * @return string[]
-     *
-     * @deprecated since Symfony 7.1, to be removed in 8.0
-     */
-    public function getAnnotatedClassesToCompile(): array
-    {
-        trigger_deprecation('symfony/http-kernel', '7.1', 'The "%s()" method is deprecated since Symfony 7.1 and will be removed in 8.0.', __METHOD__);
-
-        return $this->annotatedClasses;
-    }
-
-    /**
-     * Adds annotated classes to the class cache.
-     *
-     * @param string[] $annotatedClasses An array of class patterns
-     *
-     * @deprecated since Symfony 7.1, to be removed in 8.0
-     */
-    public function addAnnotatedClassesToCompile(array $annotatedClasses): void
-    {
-        trigger_deprecation('symfony/http-kernel', '7.1', 'The "%s()" method is deprecated since Symfony 7.1 and will be removed in 8.0.', __METHOD__);
-
-        $this->annotatedClasses = array_merge($this->annotatedClasses, $annotatedClasses);
-    }
 }